Allow to not depend on offscreen_gl_context
