+++ webrender_traits/Cargo.toml
@@ -8,6 +8,7 @@ repository = "https://github.com/servo/webrender"
 [features]
 nightly = ["euclid/unstable", "serde/unstable"]
 ipc = ["ipc-channel"]
+webgl = ["offscreen_gl_context"]
 
 [dependencies]
 app_units = "0.4"
@@ -16,7 +17,7 @@ euclid = "0.11"
 gleam = "0.4"
 heapsize = "0.3.6"
 ipc-channel = {version = "0.7", optional = true}
-offscreen_gl_context = {version = "0.8", features = ["serde"]}
+offscreen_gl_context = {version = "0.8", features = ["serde"], optional = true}
 serde = "0.9"
 serde_derive = "0.9"
